usm: http2: Add incomplete_frame_type_t enum
The new enum represents the different options for an incomlete frame.
Either we have an incomplete frame header or an incomplete frame payload.

Having also a default state 'unknown' to avoid adding bugs of relying on the default
value of the type as either incomlete frame header or an incomplete frame payload